requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
aioredis
pydantic==1.10.14
passlib[bcrypt]
python-jose[cryptography]
celery[redis]
pytest
pytest-asyncio
httpx
python-multipart
structlog 
alembic 
motor
email-validator